% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims.R
\name{mochims}
\alias{mochims}
\title{mochims}
\usage{
mochims(
  startStage = 1,
  stopStage = 5,
  base_dir = "/users/project/prj004631/afaure/DMS/Results/mochims_proj",
  output_dir = ""
)
}
\arguments{
\item{startStage}{Start at a specified analysis stage (default:2)}

\item{stopStage}{Stop at a specified analysis stage (default:7)}

\item{base_dir}{Base directory for all input files (default:NB private CRG server path; change accordingly)}

\item{output_dir}{Output directory for all output files (default:same as base_dir)}
}
\value{
Nothing
}
\description{
Main analysis script.
}
