% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims__create_dir.R
\name{mochims__create_dir}
\alias{mochims__create_dir}
\title{mochims__create_dir}
\usage{
mochims__create_dir(
  mochims_dir,
  execute = TRUE,
  message = NULL,
  overwrite_dir = TRUE
)
}
\arguments{
\item{mochims_dir}{directory path string (required)}

\item{execute}{whether or not the system command will be executed (required)}

\item{message}{message string (optional, default: NULL i.e. no message displayed)}

\item{overwrite_dir}{delete directory if already exists (optional, default: TRUE)}
}
\value{
Nothing
}
\description{
Create results folder for analysis script plots and saved objects.
}
