% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims__distance_matrix_from_PDB.R
\name{mochims__distance_matrix_from_PDB}
\alias{mochims__distance_matrix_from_PDB}
\title{mochims__distance_matrix_from_PDB}
\usage{
mochims__distance_matrix_from_PDB(
  input_file,
  residue_subset = "all",
  query_chain = "A",
  metric = "scHA"
)
}
\arguments{
\item{input_file}{path to PDB file (required)}

\item{residue_subset}{subset of residues (default:all)}

\item{query_chain}{query chain (default:A)}

\item{metric}{distance metric (default:'scHA')}
}
\value{
data.table with minimum residue distances
}
\description{
Residue distance matrix from PDB.
}
