% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims__get_reference_id.R
\name{mochims__get_reference_id}
\alias{mochims__get_reference_id}
\title{mochims__get_reference_id}
\usage{
mochims__get_reference_id(input_dt, position_offset = 0)
}
\arguments{
\item{input_dt}{data.table with id and mut_order columns (required)}

\item{position_offset}{residue position offset (default:0)}
}
\value{
Reference id with absolute amino acid residue position
}
\description{
Get reference id with absolute amino acid residue position.
}
