% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims__plot_additive_trait_2state.R
\name{mochims__plot_additive_trait_2state}
\alias{mochims__plot_additive_trait_2state}
\title{mochims__plot_additive_trait_2state}
\usage{
mochims__plot_additive_trait_2state(
  mochi_outpath,
  input_dt,
  RT = 0.001987 * (273 + 24),
  trait_name = "Folding",
  phenotype_name = "Abundance",
  report_outpath,
  dataset_name,
  colour_scheme
)
}
\arguments{
\item{mochi_outpath}{path to MoCHI thermo model fit results (required)}

\item{input_dt}{data.table with model free energy estimates (required)}

\item{RT}{constant (default:0.001987*(273+24))}

\item{trait_name}{additive trait name (default:'Folding')}

\item{phenotype_name}{phenotype name (default:'Abundance')}

\item{report_outpath}{output path for scatterplots (required)}

\item{dataset_name}{dataset name and suffix for output files (required)}

\item{colour_scheme}{colour scheme file (default:ggplot colours)}
}
\value{
Nothing
}
\description{
Plot additive trait for 2-state model.
}
