% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims__plot_additive_trait_3state.R
\name{mochims__plot_additive_trait_3state}
\alias{mochims__plot_additive_trait_3state}
\title{mochims__plot_additive_trait_3state}
\usage{
mochims__plot_additive_trait_3state(
  mochi_outpath,
  input_dt,
  RT = 0.001987 * (273 + 24),
  trait_names = c("Folding", "Binding"),
  phenotype_name = "Binding",
  report_outpath,
  dataset_name,
  colour_scheme
)
}
\arguments{
\item{mochi_outpath}{path to MoCHI thermo model fit results (required)}

\item{input_dt}{data.table with model free energy estimates (required)}

\item{RT}{constant (default:0.001987*(273+24))}

\item{trait_names}{additive trait name (default:c('Folding', 'Binding'))}

\item{phenotype_name}{phenotype name (default:'Binding')}

\item{report_outpath}{output path for scatterplots (required)}

\item{dataset_name}{dataset name and suffix for output files (required)}

\item{colour_scheme}{colour scheme file (default:ggplot colours)}
}
\value{
Nothing
}
\description{
Plot folding additive trait.
}
