% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims__predict_fitness_2state.R
\name{mochims__predict_fitness_2state}
\alias{mochims__predict_fitness_2state}
\title{mochims__predict_fitness_2state}
\usage{
mochims__predict_fitness_2state(
  mochi_outpath,
  phenotype_name,
  energy,
  RT = 0.001987 * (273 + 24)
)
}
\arguments{
\item{mochi_outpath}{path to MoCHI thermo model fit results (required)}

\item{phenotype_name}{phenotype name (required)}

\item{energy}{energies for 2-state model (required)}

\item{RT}{constant (default:0.001987*(273+24))}
}
\value{
A list with fitness predictions
}
\description{
Predict fitness from energies for 2-state model.
}
