% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims__predict_fitness_3state.R
\name{mochims__predict_fitness_3state}
\alias{mochims__predict_fitness_3state}
\title{mochims__predict_fitness_3state}
\usage{
mochims__predict_fitness_3state(
  mochi_outpath,
  phenotype_name,
  folding_energy,
  binding_energy,
  RT = 0.001987 * (273 + 24)
)
}
\arguments{
\item{mochi_outpath}{path to MoCHI thermo model fit results (required)}

\item{phenotype_name}{phenotype name (required)}

\item{folding_energy}{folding energies for 3-state model (required)}

\item{binding_energy}{binding energies for 3-state model (required for binding fitness prediction)}

\item{RT}{constant (default:0.001987*(273+24))}
}
\value{
A list with folding and binding fitness predictions
}
\description{
Predict fitness from folding and binding energies for 3-state model.
}
