% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims_coefficient_size_distribution.R
\name{mochims_coefficient_size_distribution}
\alias{mochims_coefficient_size_distribution}
\title{mochims_coefficient_size_distribution}
\usage{
mochims_coefficient_size_distribution(input_file, outpath, colour_scheme)
}
\arguments{
\item{input_file}{path to MoCHI thermo model fit results (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}
}
\value{
Nothing
}
\description{
Coupling scatterplots.
}
