% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mochims_coefficient_size_distribution_datasets.R
\name{mochims_coefficient_size_distribution_datasets}
\alias{mochims_coefficient_size_distribution_datasets}
\title{mochims_coefficient_size_distribution_datasets}
\usage{
mochims_coefficient_size_distribution_datasets(
  dataset_names,
  validation_list,
  base_dir,
  output_dir,
  stagenum,
  colour_scheme,
  execute = TRUE
)
}
\arguments{
\item{dataset_names}{character vector of dataset names (required)}

\item{base_dir}{Base directory (required)}

\item{output_dir}{Output directory (required)}

\item{stagenum}{stage number (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Coefficient size distribution plots for multiple datasets.
}
