% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims_coupling_scatter.R
\name{mochims_coupling_scatter}
\alias{mochims_coupling_scatter}
\title{mochims_coupling_scatter}
\usage{
mochims_coupling_scatter(
  input_file,
  validation_couplings,
  outpath,
  colour_scheme
)
}
\arguments{
\item{input_file}{path to MoCHI thermo model fit results (required)}

\item{validation_couplings}{validation coupling ids (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}
}
\value{
Nothing
}
\description{
Coupling scatterplots.
}
