% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims_coupling_scatter_datasets.R
\name{mochims_coupling_scatter_datasets}
\alias{mochims_coupling_scatter_datasets}
\title{mochims_coupling_scatter_datasets}
\usage{
mochims_coupling_scatter_datasets(
  dataset_names,
  validation_list,
  base_dir,
  output_dir,
  stagenum,
  colour_scheme,
  execute = TRUE
)
}
\arguments{
\item{dataset_names}{character vector of dataset names (required)}

\item{validation_list}{list of validation couplings (required)}

\item{base_dir}{Base directory (required)}

\item{output_dir}{Output directory (required)}

\item{stagenum}{stage number (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Coupling scatterplots for multiple datasets.
}
