% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims_position_class_violins.R
\name{mochims_position_class_violins}
\alias{mochims_position_class_violins}
\title{mochims_position_class_violins}
\usage{
mochims_position_class_violins(
  input_file,
  annotation_file,
  outpath,
  colour_scheme
)
}
\arguments{
\item{input_file}{path to MoCHI thermo model fit results (required)}

\item{annotation_file}{path to MoCHI thermo model fit results (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}
}
\value{
Nothing
}
\description{
Coupling scatterplots.
}
