% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims_position_class_violins_datasets.R
\name{mochims_position_class_violins_datasets}
\alias{mochims_position_class_violins_datasets}
\title{mochims_position_class_violins_datasets}
\usage{
mochims_position_class_violins_datasets(
  dataset_names,
  annotations,
  base_dir,
  output_dir,
  stagenum,
  colour_scheme,
  execute = TRUE
)
}
\arguments{
\item{dataset_names}{character vector of dataset names (required)}

\item{annotations}{Annotations directory (required)}

\item{base_dir}{Base directory (required)}

\item{output_dir}{Output directory (required)}

\item{stagenum}{stage number (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Coupling scatterplots for multiple datasets.
}
