% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mochims_structure_metrics.R
\name{mochims_structure_metrics}
\alias{mochims_structure_metrics}
\title{mochims_structure_metrics}
\usage{
mochims_structure_metrics(input_file, outpath, pdb_file, execute = TRUE)
}
\arguments{
\item{input_file}{path to MoCHI thermo model fit results (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{pdb_file}{path to PDB file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Add 3D structure metrics.
}
